Refactor folder, map component to centralize logic
