+++ nimbus-web/components/MapPageComponents/PlanTab/Folders/FullScreenPlan.tsx
