+++ nimbus-web/pages/_app.tsx
@@ -2,7 +2,7 @@ import "@/styles/globals.css";
 import type { AppProps } from "next/app";
 import { UserProvider } from "@auth0/nextjs-auth0/client";
 import Layout from "@/components/Layout";
-import { PlanTabProvider } from "@/components/PlanTab/PlanTabContext";
+import { PlanTabProvider } from "@/components/MapPageComponents/PlanTab/PlanTabContext";
 import { MapProvider } from "react-map-gl";
 
 export default function App({ Component, pageProps }: AppProps) {
